fixed cart

+++ src/models/cartDao.js
@@ -1,6 +1,7 @@
 const appDataSource = require('./appDataSource');
 const { DatabaseError } = require('../middlewares/error');
 const { orderStatusEnum } = require('../middlewares/enums');
+const { AddOrUpdate } = require('./conditionMake');
 
 const createCartItem = async ({ userId, productId, quantity }) => {
   try {
@@ -8,9 +9,11 @@ const createCartItem = async ({ userId, productId, quantity }) => {
       'SELECT * FROM product where product.id =?',
       [productId]
     );
+
     if (product <= 0) {
       throw new DatabaseError('Invalid product');
     }
+
     await appDataSource.query(
       `INSERT INTO cart(user_id, product_items, quantity)
           VALUES (?, ?, ?)`,
@@ -28,6 +31,7 @@ const existCartItem = async (userId, productId) => {
     `SELECT * FROM cart WHERE product_items = ? AND user_id =?`,
     [productId, userId]
   );
+
   const cartArray = Array.isArray(cart) ? cart : [cart];
   return cartArray;
 };
@@ -96,17 +100,20 @@ const getCartList = async (userId) => {
   return updatedLists;
 };
 
-const updateCartItemQuantity = async ({ quantity, userId, productId }) => {
+const updateCartItemQuantity = async ({
+  quantity,
+  userId,
+  productId,
+  addorupdatestatusEnum,
+}) => {
+  let addorupdate = new AddOrUpdate(quantity, userId, productId);
+  if (!addorupdatestatusEnum) addorupdate = addorupdate.add();
+  if (addorupdatestatusEnum) addorupdate = addorupdate.update();
   const queryRunner = appDataSource.createQueryRunner();
   await queryRunner.connect();
   await queryRunner.startTransaction();
   try {
-    await queryRunner.query(
-      `UPDATE cart
-       SET quantity = ?
-       WHERE cart.user_id = ? AND cart.product_items = ? `,
-      [quantity, userId, productId]
-    );
+    await queryRunner.query(addorupdate, [quantity, userId, productId]);
 
     const [updatedCartItem] = await queryRunner.query(
       `SELECT
@@ -127,7 +134,7 @@ const updateCartItemQuantity = async ({ quantity, userId, productId }) => {
   }
 };
 
-const deleteCart = async ({ userId, productId }) => {
+const deleteCart = async (userId, productId) => {
   await appDataSource.query(
     `
         DELETE
@@ -137,59 +144,40 @@ const deleteCart = async ({ userId, productId }) => {
     [userId, productId]
   );
 
-  const pendingPayment = orderStatusEnum.PENDING_PAYMENT;
-
-  const order = await appDataSource.query(
+  const existingOrder = await appDataSource.query(
     `SELECT
+     *
+    FROM
+      orders
+    WHERE
+      user_id =?`,
+    [userId]
+  );
+
+  const existingOrderArray = Array.isArray ? existingOrder : [existingOrder];
+
+  if (existingOrderArray.length > 0) {
+    const pendingPayment = orderStatusEnum.PENDING_PAYMENT;
+
+    const order = await appDataSource.query(
+      `SELECT
       id
     FROM
       orders
     WHERE
       user_id =? AND order_status_id =?
     `,
-    [userId, pendingPayment]
-  );
-  const orderId = order[0].id;
+      [userId, pendingPayment]
+    );
+    const orderId = order[0].id;
 
-  await appDataSource.query(
-    `DELETE FROM order_item WHERE order_id = ? AND product_id = ?`,
-    [orderId, productId]
-  );
+    await appDataSource.query(
+      `DELETE FROM order_item WHERE order_id = ? AND product_id = ?`,
+      [orderId, productId]
+    );
+  }
 
   return 'cartDeleted';
-};
-
-const addCartItemQuantity = async ({ userId, productId, quantity }) => {
-  const queryRunner = appDataSource.createQueryRunner();
-
-  await queryRunner.connect();
-  await queryRunner.startTransaction();
-
-  try {
-    await queryRunner.query(
-      `UPDATE cart
-       SET quantity = quantity+ ?
-       WHERE cart.user_id = ? AND cart.product_items = ? `,
-      [quantity, userId, productId]
-    );
-
-    const [updatedCartItem] = await queryRunner.query(
-      `SELECT
-         id, user_id, product_items, quantity
-       FROM cart
-       WHERE cart.user_id = ? AND cart.product_items = ?`,
-      [userId, productId]
-    );
-
-    await queryRunner.commitTransaction();
-
-    return { message: 'cartItem quantity updated', updatedCartItem };
-  } catch (error) {
-    await queryRunner.rollbackTransaction();
-    throw new DatabaseError('failed to update cart item quantity');
-  } finally {
-    await queryRunner.release();
-  }
 };
 
 module.exports = {
@@ -199,6 +187,5 @@ module.exports = {
   updateCartItemQuantity,
   deleteCart,
   findMatchedProductId,
-  addCartItemQuantity,
   existCartItem,
 };
